# oas/odin-schema.yml
openapi: '3.0.2'
info:
  title: API Title
  version: '0.0.1'
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /jobs/execution/:
    get:
      summary: Get all execution jobs
      description: Get all execution jobs
      operationId: getAllExecutionJobs
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - jobs
                  - pagination
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /executions/:
    get:
      summary: Get all executions
      description: Get all executions
      operationId: getAllExecutions
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - executions
                  - pagination
                properties:
                  executions:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExecutionResult'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /jobs/{JobId}/executions/:
    get:
      summary: Get executions of given job
      description: Get executions of given job
      operationId: getExecutionsForJob
      parameters:
        - name: JobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - executions
                  - pagination
                properties:
                  executions:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExecutionResult'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Execution job not found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /executions/{execId}/:
    get:
      summary: Get execution result by id
      description: Get execution result by id
      operationId: getExecutionResultById
      parameters:
        - name: execId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResult'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Execution result not found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /executions/jobs/{JobId}/:
    get:
      summary: Get execution job
      description: Get execution job
      operationId: getExecutionJobById
      parameters:
        - name: JobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Execution job not found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Cancel Execution Job
      description: Cancel Execution Job
      operationId: cancelExecutionJob
      parameters:
        - name: JobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete execution job
      description: Delete execution job
      operationId: deleteExecutionJob
      parameters:
        - name: JobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Execution job not found
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /executions/workers:
    get:
      summary: Get all execution workers
      description: Get all execution workers
      operationId: getExecutionWorkers
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - workers
                  - pagination
                properties:
                  workers:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExecutionWorker'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /execution/config/:
    get:
      summary: Get execution config
      description: Get execution config
      operationId: getExecutionConfig
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionConfig'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
components:
  parameters:
    PageNumber:
      name: page
      in: query
      description: The page number to retrieve.
      required: false
      schema:
        type: integer
        format: int32
        default: 0
    PageSize:
      name: pageSize
      in: query
      description: The number of items per page.
      required: false
      schema:
        type: integer
        format: int32
        default: 20
  schemas:
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    Job:
      type: object
      required:
        - jobId
        - script
        - flake
        - created_at
      properties:
        jobId:
          type: integer
          format: int64
        script:
          type: string
        flake:
          type: string
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ExecutionResult:
      allOf:
        - $ref: '#/components/schemas/Job'
        - type: object
          required:
            - execId
            - script
            - flake
            - args
            - started_at
            - finished_at
            - exec_logs
          properties:
            execId:
              type: integer
              format: int64
            script:
              type: string
            flake:
              type: string
            args:
              type: string
            started_at:
              type: string
              format: date-time
            finished_at:
              type: string
              format: date-time
            exec_logs:
              type: string
            nix_logs:
              type: string
...
    ExecutionWorker:
      type: object
      required:
        - id
        - name
        - status
        - created_at
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        created_at:
          type: string
          format: date-time
        status:
          type: string
        updated_at:
          type: string
          format: date-time
    ExecutionConfig:
      type: object
      required:
        - ODIN_WORKER_PROVIDER
        - ODIN_WORKER_CONCURRENCY
        - ODIN_WORKER_BUFFER_SIZE
        - ODIN_WORKER_TASK_TIMEOUT
        - ODIN_WORKER_POLL_FREQ
        - ODIN_WORKER_RUNTIME
        - ODIN_LOG_LEVEL
      properties:
        ODIN_WORKER_PROVIDER:
          type: string
          description: Represents the worker provider.
        ODIN_WORKER_CONCURRENCY:
          type: integer
          format: int32
          description: Represents the concurrency level for the worker.
        ODIN_WORKER_BUFFER_SIZE:
          type: integer
          format: int32
          description: Represents the buffer size for the worker.
        ODIN_WORKER_TASK_TIMEOUT:
          type: integer
          description: Represents the task timeout.
        ODIN_WORKER_POLL_FREQ:
          type: integer
          description: Represents the polling frequency for the worker in seconds.
        ODIN_WORKER_RUNTIME:
          type: string
          description: Represents the runtime for the worker in seconds.
        ODIN_LOG_LEVEL:
          type: string
          description: Represents the log level.
        ODIN_SYSTEM_PROVIDER_BASE_DIR:
          type: string
          description: Represents the base directory for the system provider.
        ODIN_SYSTEM_PROVIDER_CLEAN_UP:
          type: boolean
          description: Represents whether to clean up directories created by the system provider.
    PaginationResponse:
      type: object
      required:
        - total
        - limit
        - offset
      properties:
        total:
          type: integer
          format: int64
          description: Represents the total number of items.
        pages:
          type: integer
          format: int32
          description: Represents the total number of pages.
        page:
          type: integer
          format: int32
          description: Represents the current page.
        limit:
          type: integer
          format: int32
          description: Represents the number of items per page.
        next:
          type: string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []